+++ packages/react-openapi/src/index.ts
@@ -1,4 +1,6 @@
 export * from './schemas';
+export * from './formatOpenAPIMethod';
+export * from './OpenAPIMethodBadge';
 export * from './OpenAPIOperation';
 export * from './OpenAPIWebhook';
 export * from './OpenAPIOperationContext';